feat: 取消user表主键，修改其他id为bigint

+++ featherwit-blogger/model/request/blog.go
@@ -12,11 +12,11 @@ type SearchBlogRequest struct {
 }
 
 type BlogIdRequest struct {
-	ID int `json:"id" uri:"id"`
+	ID int64 `json:"id" uri:"id"`
 }
 
 type UpdateBlogRequest struct {
-	ID      int    `json:"id" bind:"required"`
+	ID      int64  `json:"id" bind:"required"`
 	Title   string `json:"title"`
 	Content string `json:"content"`
 }
